Stylistic Changes and issue with overwrite
hdu.writeto() wanted the old clobber=True, not overwrite=True. This change was
needed on my machine, but won't be on others. The best action would be to check
for astropy version and use the old version if you have to.
Also, hdu.writeto verify_check was set to ignore.

Moved ExtractSpectra main method to a main method. This will make it possible
to import to other programs without needing to run the code.
Trimmed line lengths to 160-character cap. 80 is the nominal line cap in python,
but it isn't my code, so I'm not going to set style guidelines.