Added Interactive mode, fixed out-of-range issues
CrossCor now defaults to running through an entire mask without prompting
for any user input; the user is then expected to go back and manually select
which images they want to look at again. This is admittedly done sloppily;
it generates the default GUI, but then immediately triggers Bsave().

There was an issue with avoiding zeroes and going out of range on wavelength.
Kelson fixed that. Before, parts of the fit would not appear on the GUI as
anything other than a flat line for out-of-range parts.

Finally, the main class was modified to not call an array of im(k)s. This
prevents the computer from having too many open processes, solving an issue
I was having.