Usability upgrades, including arg handling, GUI flagging, and z_err finding
1. Args are now passed in the traditional way (-w, --words), which lets us
use multiple command line arguments at once. New CLAs include -e, for
assuming emission lines are present, and -b, for a set of objects known to
be bad, and flagged accordingly a priori (for batch, uninteractive jobs).
2. When Star or Bad is marked, those buttons now turn red; switching them
back turns the buttons gray again. This gives a quick visual indicator
when scanning through multiple output cc_{}.png files that this is not
supposed to be a good object.
3. There is an issue where super-pointy spikes in cc cause the root
finding to break. We've done some work on mitigating this issue,
including switching the interp1d to using bsplines, putting a cap
on dz (to roughly the pixel size), and limiting how far the routine
can run. More work may still happen.